refactor - using Questions.js page instead of Forums.js, also added index filter to display results by category
